Document pattern emission
